rename/addgamelist to show game list screen
